Add files via upload
Now, Thermoemf (E) varies with Temperature (T) as E=A*T+B*T^2,where A and B are obtained using the Supervised Learning Model. Now, using this polynomial as the fitting model simply underfits the data ,ie,high bias is encountered. Hence, we used a 4th order polynomial model to fit the dataset just right. In this way,we have obtained desirable results in this project.